feat(auth): implement JWT authentication and user token validation
- Added utility to create and decode JWTs
- Implemented OAuth2PasswordBearer dependency
- Secured routes using token-based authentication
