+++ server/auth/hash_utils.py
@@ -1,9 +1,4 @@
-from jose import jwt
-from datetime import datetime, timedelta
 from passlib.context import CryptContext
-
-from config import ACCESS_TOKEN_EXPIRE_MINUTES, ALGORITHM, SECRET_KEY
-
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated = "auto")
 
 def hash_password(password: str) -> str:
@@ -12,9 +7,4 @@ def hash_password(password: str) -> str:
 def verify_password(plain_pass: str, hash_pass: str) -> bool:
     return pwd_context.verify(plain_pass,hash_pass)
 
-def create_access_token(data: dict):
-    to_encode = data.copy()
-    expire = datetime.utcnow() + timedelta(minutes= ACCESS_TOKEN_EXPIRE_MINUTES)
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
+
